upgrading the project to spring 5 and refactored it to use gemfire-spring-boot-starter
