+++ manifest.yml
@@ -6,5 +6,5 @@ applications:
   path: target/cache-demo-0.0.1-SNAPSHOT.jar
   buildpack: java_buildpack_offline
   services:
-  - test-pcc
-  - syf-mysql
+  - dev-cluster
+  - mysql-dev
