+++ pom.xml
@@ -11,31 +11,31 @@
 	<name>cache-demo</name>
 	<description>Demo project for Spring Boot</description>
 
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+        <java.version>1.8</java.version>
+    </properties>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.4.2.RELEASE</version>
+		<version>2.0.2.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-	</properties>
-
 	<dependencies>
-		<!-- Spring Data Geode -->
-		<dependency>
-			<groupId>org.springframework.data</groupId>
-			<artifactId>spring-data-geode</artifactId>
-			<version>1.0.0.INCUBATING-RELEASE</version>
-		</dependency>
 		
 		<!-- Spring MVC -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-starter-logging</artifactId>
+                </exclusion>
+            </exclusions>
 		</dependency>
 		
 		<!-- jfairy Fake Data Generator: https://github.com/Codearte/jfairy-->
@@ -49,25 +49,47 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>org.springframework.boot</groupId>
+                    <artifactId>spring-boot-starter-logging</artifactId>
+                </exclusion>
+            </exclusions>
 		</dependency>
-		
-		<!-- Spring Cloud Connector for Cloud Foundry -->
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-cloudfoundry-connector</artifactId>
-		</dependency>
-		
-		<!-- Spring Cloud GemFire Connector: https://github.com/pivotal-cf/spring-cloud-gemfire-connector/tree/v1.1.0 -->
-		<dependency>
-			<groupId>io.pivotal.spring.cloud</groupId>
-			<artifactId>spring-cloud-gemfire-cloudfoundry-connector</artifactId>
-			<version>1.1.0.RELEASE</version>
-		</dependency>
-		<dependency>
-			<groupId>io.pivotal.spring.cloud</groupId>
-			<artifactId>spring-cloud-gemfire-spring-connector</artifactId>
-			<version>1.1.0.RELEASE</version>
-		</dependency>
+
+        <!-- Spring Boot Data Geode -->
+        <dependency>
+            <groupId>org.springframework.geode</groupId>
+            <artifactId>gemfire-spring-boot-starter</artifactId>
+            <version>1.0.0.BUILD-SNAPSHOT</version>
+        </dependency>
+
+        <!-- Spring Cloud Connector for Cloud Foundry -->
+
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-spring-service-connector</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-cloudfoundry-connector</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.apache.logging.log4j</groupId>
+            <artifactId>log4j-slf4j-impl</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jcl-over-slf4j</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>jul-to-slf4j</artifactId>
+        </dependency>
+
 		
 		<!-- HyperSQL DB -->
 		<dependency>
@@ -89,6 +111,11 @@
 			<name>Spring Repository</name>
 			<url>http://repo.spring.io/libs-release</url>
 		</repository>
+        <repository>
+            <id>spring-snapshot</id>
+            <name>Spring Maven RELEASE Repository</name>
+            <url>http://repo.spring.io/libs-snapshot</url>
+        </repository>
 	</repositories>
 
 	<build>
