+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
 spring.profiles.active=cloud
+management.endpoint.env.enabled=true
+management.endpoints.web.exposure.include=env
